requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.11.7
async-timeout==4.0.2
autopep8==1.6.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bitrix24-python-sdk==1.0.1
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cssselect2==0.6.0
Deprecated==1.2.13
dill==0.3.5.1
Django==4.0.6
django-admin-sortable==2.3
django-admin-tools==0.9.2
django-ajax-selects==2.2.0
django-cacheops==6.1
django-ckeditor==6.4.2
django-colorfield==0.7.2
django-debug-toolbar==3.5.0
django-elasticsearch-dsl==7.2.2
django-filer==2.2.1
django-js-asset==2.0.0
django-light==0.1.0.post3
django-mptt==0.13.4
django-mptt-admin==2.2.0
django-polymorphic==3.1.0
django-redis==5.2.0
easy-thumbnails==2.8.2
EasyProcess==1.1
elastic-transport==8.1.2
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
funcy==1.17
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
multidimensional-urlencode==0.0.4
olefile==0.46
packaging==21.3
pdfkit==1.0.0
pep8==1.7.1
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycodestyle==2.9.0
pylint==2.14.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytils==0.4.1
pytz==2022.1
PyVirtualDisplay==3.0
redis==4.3.4
reportlab==3.6.11
requests==2.28.1
sentry-sdk==1.16.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
svglib==1.3.0
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
Unidecode==1.1.2
urllib3==1.26.11
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.1
django-import-export
gunicorn
